fop upgrade with png support. build tool is now gradle. plugin works in jetty.
