+++ Extensions.xml
@@ -8,11 +8,11 @@
 		<initParam name="className">com.percussion.pso.fop.FopAssembler</initParam>
 		<initParam name="com.percussion.extension.assembly.fileSuffix">.vm</initParam>
 		<initParam name="com.percussion.extension.assembly.autoReload">yes</initParam>
-		<initParam name="com.percussion.pso.fop.SupportedMimeTypes">application/pdf,application/postscript,application/x-pcl,text/plain,text/richtext,text/rtf,application/rtf</initParam>
+		<initParam name="com.percussion.pso.fop.SupportedMimeTypes">application/pdf,application/postscript,application/x-pcl,text/plain,text/richtext,text/rtf,application/rtf,image/png</initParam>
 		<interface name="com.percussion.services.assembly.IPSAssembler"/>		
 		<suppliedResources/>
 	</Extension>
-		<Extension name="session" context="global/percussion/system" handler="Java" categorystring="jexl" deprecated="no">
+	<Extension name="session" context="global/percussion/system" handler="Java" categorystring="jexl" deprecated="no">
 		<initParam name="com.percussion.user.description">Jexl extension for certain session data required for authentication.</initParam>
 		<initParam name="com.percussion.extension.version">1</initParam>
 		<initParam name="className">com.percussion.services.assembly.jexl.PSSessionUtils</initParam>
