--- build.number
@@ -1,3 +1,0 @@
-#Build Number for ANT. Do not edit!
-#Wed May 09 16:35:54 EDT 2007
-build.number=4
