--- deploy.xml
@@ -1,65 +1,0 @@
-<project name="Rhino Dynamic Delivery Deploy" default="dist" basedir=".">
-	<description>
-       Deployment Script Rhino Dynamic Delivery 
-    </description>
-    
-    <property environment="env"/>
-	<!-- base directory of Rhythmyx installation --> 
-	<property name="rhino.home" location="${env.RHYTHMYX_HOME}"/>
-	<!-- name of our output jar --> 
-	<property name="jarname" value="dynapub-server.jar" />	
-	<property name="pubjarname" value="dynapub-publisher.jar" /> 
-	
-	<property name="lib" location="lib"/>
-	
-	<property name="rhino.deploy" value="${rhino.home}/AppServer/server/rx/deploy" /> 
-	<property name="rhino.rxapp"  value="${rhino.deploy}/rxapp.ear/rxapp.war" /> 
-	<property name="rhino.rxpub"  value="${rhino.deploy}/RxServices.war"/> 
-	<property name="rhino.rxapp.lib"    value="${rhino.rxapp}/WEB-INF/lib"/>
-	<property name="rhino.rxapp.spring-config"
-	value="${rhino.rxapp}/WEB-INF/config/user/spring" />
-	
-	<property name="rhino.rxpub.lib" value="${rhino.rxpub}/WEB-INF/lib" />
-	<property name="rhino.rxpub.spring-config" value="${rhino.rxpub}/WEB-INF/config/spring" /> 
-		
-	
-	<target name="installExtensions">
-	    <fail unless="rhino.home" message="You must set the RHYTHMYX_HOME environment variable" />
-		<java classname="com.percussion.util.PSExtensionInstallTool" fork="true">
-		    <arg value="${rhino.home}" />
-			<arg value="${basedir}" />
-			<classpath>
-				<fileset dir="${rhino.rxapp.lib}"/>
-			    <fileset dir="${rhino.home}/AppServer/lib/endorsed" />
-				<fileset dir="${rhino.home}/AppServer/server/rx/lib" />
-			</classpath>			
-		</java>     
-	</target>
-	
-	
-	<target name="deployServer" description="deployment of server jars and config">		
-	       <fail unless="rhino.rxapp.lib" message="You must set the RHYTHMYX_HOME environment variable" />		
-		    <copy todir="${rhino.rxapp.lib}" verbose="true"> 
-		        <fileset dir="${lib}"/>
-		    </copy>
-		    <chmod perm="755">
-                <fileset dir="${rhino.rxapp.lib}" />
-			</chmod>	
-		    <copy todir="${rhino.home}/rxconfig" verbose="true">
-		    	<fileset dir="rxconfig" />
-		    </copy>
-	</target>
-
-	<target name="deploySamples" description="Deploy Sample code">
-	       <fail unless="rhino.rxapp.lib" message="You must set the RHYTHMYX_HOME environment variable" />		
-		   <copy todir="${rhino.home}/rx_resources" verbose="true">
-		      <fileset dir="Samples" includes="images/** stylesheets/** " />
-		   </copy>  
-	</target>
-	
-	<target name="dist" description="Default target deploys all">
-		<fail unless="rhino.home" message="You must set the RHYTHMYX_HOME environment variable" />		
-		<antcall target="installExtensions"/>
-		<antcall target="deployServer" />
-	</target>
-</project>
