+++ rxconfig/Server/fop.config
@@ -16,16 +16,16 @@ the location of this file.
 
   <!-- Base URL for resolving relative URLs -->
   <base>.</base>
-  
+
   <!-- Source resolution in dpi (dots/pixels per inch) for determining the size of pixels in SVG and bitmap images, default: 72dpi -->
   <source-resolution>72</source-resolution>
   <!-- Target resolution in dpi (dots/pixels per inch) for specifying the target resolution for generated bitmaps, default: 72dpi -->
   <target-resolution>72</target-resolution>
-  
+
   <!-- Default page-height and page-width, in case
        value is specified as auto -->
   <default-page-settings height="11in" width="8.26in"/>
-  
+
   <!-- Information for specific renderers -->
   <!-- Uses renderer mime type for renderers -->
   <renderers>
@@ -33,11 +33,11 @@ the location of this file.
       <filterList>
         <!-- provides compression using zlib flate (default is on) -->
         <value>flate</value>
-  
+
         <!-- encodes binary data into printable ascii characters (default off)
              This provides about a 4:5 expansion of data size -->
         <!-- <value>ascii-85</value> -->
-  
+
         <!-- encodes binary data with hex representation (default off)
              This filter is not recommended as it doubles the data size -->
         <!-- <value>ascii-hex</value> -->
@@ -66,7 +66,6 @@ the location of this file.
                           | 500 | 600 | 700 | 800 | 900
         (normal = 400, bold = 700)
         -->
-
         <!--
         <font metrics-url="arial.xml" kerning="yes" embed-url="arial.ttf">
           <font-triplet name="Arial" style="normal" weight="normal"/>
@@ -76,14 +75,17 @@ the location of this file.
           <font-triplet name="Arial" style="normal" weight="bold"/>
           <font-triplet name="ArialMT" style="normal" weight="bold"/>
         </font>
-        -->
-          <font metrics-url="file:///D:/usergroup-fonts/arial.xml" kerning="yes" embed-url="file:///D:/usergroup-fonts/ARIAL.TTF">
-	    <font-triplet name="Arial" style="normal" weight="normal"/>
-        </font>      
+        <font metrics-url="file:///D:/usergroup-fonts/arial.xml" kerning="yes" embed-url="file:///D:/usergroup-fonts/ARIAL.TTF">
+	      <font-triplet name="Arial" style="normal" weight="normal"/>
+        </font>
         <font metrics-url="file:///D:/usergroup-fonts/baskerville.xml" kerning="yes" embed-url="file:///D:/usergroup-fonts/BASKVILL.TTF">
 	    <font-triplet name="BaskervilleOldFace" style="normal" weight="normal"/>
 	    <font-triplet name="Baskerville Old Face" style="normal" weight="normal"/>
         </font>
+        <font kerning="yes" embed-url="/usr/share/fonts/truetype/msttcorefonts/Arial.ttf" encoding-mode="auto">
+          <font-triplet name="Arial" style="normal" weight="normal" />
+        </font>
+        -->
       </fonts>
 
       <!-- This option lets you specify additional options on an XML handler -->
@@ -96,7 +98,7 @@ the location of this file.
     <renderer mime="application/postscript">
       <!-- This option forces the PS renderer to rotate landscape pages -->
       <!--auto-rotate-landscape>true</auto-rotate-landscape-->
-      
+
       <!-- This option lets you specify additional options on an XML handler -->
       <!--xml-handler namespace="http://www.w3.org/2000/svg">
         <stroke-text>false</stroke-text>
@@ -130,6 +132,15 @@ the location of this file.
 
     <renderer mime="text/plain">
       <pageSize columns="80"/>
+    </renderer>
+
+    <renderer mime="image/png">
+      <!--<color-mode>rgba</color-mode>-->
+      <!--<transparent-page-background>true</transparent-page-background>-->
+      <!--<background-color>white</background-color>-->
+      <!--<anti-aliasing>true</anti-aliasing>-->
+      <!--<rendering>quality</rendering>-->
+      <!--<fonts>described elsewhere</fonts>-->
     </renderer>
 
   </renderers>
